+++ CHANGELOG.md
@@ -4,8 +4,16 @@
 
 ## Unreleased
 
-* Enh: Use `hatch_vcs` for dynamic versioning (@lwasser, #82)
-* Fix: migrate to pytest for tests and setup hatch scripts (#89, @lwasser)
+- Enh: Use `hatch_vcs` for dynamic versioning (@lwasser, #82)
+- Fix: migrate to pytest for tests and setup hatch scripts (#89, @lwasser)
+- Add: Partner support to package (#92, @lwasser)
+- Add: Code coverage setup (#97, @lwasser)
+- Fix: Add validation step to categories attr in pydantic model (#105, @lwasser)
 
-## v.0.2
-lots of stuff here... :) none of it documented
+## [v0.15](https://github.com/pyOpenSci/pyosMeta/releases/tag/v0.15)
+
+2023-12-20 - This release was tagged in git and GitHub but not released to PyPI.
+
+## [v.0.2](https://pypi.org/project/pyosmeta/0.2)
+
+2023-08-17 - Initial release to PyPI.